+++ scg-impl/src/main/java/com/dwarfeng/scg/impl/bean/entity/HibernateNodeVariable.java
@@ -15,7 +15,8 @@ public class HibernateNodeVariable implements Bean {
 
     private static final long serialVersionUID = 20730736142239734L;
 
-    // -----------------------------------------------------------主键-----------------------------------------------------------
+    // region 主键
+
     @Id
     @Column(name = "scg_setting_id", length = Constraints.LENGTH_ID, nullable = false)
     private String scgSettingId;
@@ -28,7 +29,10 @@ public class HibernateNodeVariable implements Bean {
     @Column(name = "variable_id", length = Constraints.LENGTH_ID, nullable = false)
     private String variableId;
 
-    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    // endregion
+
+    // region 主属性字段
+
     @Column(name = "string_value", columnDefinition = "TEXT")
     private String stringValue;
 
@@ -52,17 +56,23 @@ public class HibernateNodeVariable implements Bean {
     @Column(name = "last_updated_date")
     private Date lastUpdatedDate;
 
-    // -----------------------------------------------------------多对一-----------------------------------------------------------
+    // endregion
+
+    // region 多对一
+
     @ManyToOne(targetEntity = HibernateScgSetting.class)
     @JoinColumns({ //
             @JoinColumn(name = "scg_setting_id", referencedColumnName = "id", insertable = false, updatable = false), //
     })
     private HibernateScgSetting scgSetting;
 
+    // endregion
+
     public HibernateNodeVariable() {
     }
 
-    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
+    // region 映射用属性区
+
     public HibernateNodeVariableKey getKey() {
         return new HibernateNodeVariableKey(scgSettingId, deviceId, variableId);
     }
@@ -79,7 +89,10 @@ public class HibernateNodeVariable implements Bean {
         }
     }
 
-    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
+    // endregion
+
+    // region 常规属性区
+
     public String getScgSettingId() {
         return scgSettingId;
     }
@@ -167,6 +180,8 @@ public class HibernateNodeVariable implements Bean {
     public void setScgSetting(HibernateScgSetting scgSetting) {
         this.scgSetting = scgSetting;
     }
+
+    // endregion
 
     @Override
     public String toString() {